# Part of #2088: Speedup bigint_from_uint_arrays