Optimize the uint->bigint copy by limiting temporary array creation and avoiding what was effectively an array-of-dist-arrays reduction. For max_bits, optimize modding by localizing the `modBy` bigint.

On 16-node-cs-hdr, `bigint_from_uint_arrays` is over 100x faster when `max_bits` isn't set, and nearly 1000x faster when it is:

```python

import argparse, math, time
import arkouda as ak

parser = argparse.ArgumentParser("bigint_from_uint_arrays perf test")
parser.add_argument("hostname")
parser.add_argument("port", type=int)
parser.add_argument("--size", type=int, default=2**20)
parser.add_argument("--max-bits", type=int, default=-1)
args = parser.parse_args()

ak.connect(args.hostname, args.port)

a = ak.randint(0, 2**32, args.size, dtype=ak.uint64)
b = ak.randint(0, 2**32, args.size, dtype=ak.uint64)

start = time.time()
ba = ak.bigint_from_uint_arrays([a, b], max_bits=args.max_bits)
end = time.time()
print("time: {:.2f} sec".format(end-start))
```

```
runClient -nl 16 ./uint_to_bigint_perf.py --max-bits={-1,64} --size=$((2**{26,30}))
```

For `--size=$((2**26))` (1G, 64M/node)

| config      | before   | after |
| ----------- | -------: | ----: |
| max_bits=-1 |   3.23s  | 0.02s |
| max_bits=64 | 235.80s  | 0.03s |

For `--size=$((2**30))` (16G, 1G/node)

| config      | before   | after |
| ----------- | -------: | ----: |
| max_bits=-1 |  57.78s  | 0.12s |
| max_bits=64 |     N/A  | 0.19s |

As future work, this could be further optimized by copying the uint arrays directly into the bigint limbs and masking instead of modding.

Part of #2088